Changed how IfMatch handles a request
Had to change how the request lifecycle is handled. For some reason calling $next($request) will result in $next using an empty (or the same request as made in the middleware) request and this will result in your request failing. Now not using the next closure but instead setting the x-from-middleware header and handling the whole request again, this time skipping our ifmatch middleware. This seems suboptimal.